need to submit!

+++ server.js
@@ -14,7 +14,7 @@ const createAuthRoutes = require('./lib/auth/create-auth-routes');
 const authRoutes = createAuthRoutes({
     selectUser(email) {
         return client.query(`
-            SELECT id, email, has, display_name as "displayName
+            SELECT id, email, hash, display_name as "displayName"
             FROM users
             WHERE email = $1;
         `,
@@ -23,7 +23,7 @@ const authRoutes = createAuthRoutes({
     },
     insertUser(user, hash) {
         return client.query(`
-            INSERT into users (email, has, display_name)
+            INSERT into users (email, hash, display_name)
             VALUES ($1, $2, $3)
             RETURNING id, email, display_name as "displayName";
         `,
@@ -41,7 +41,7 @@ app.use(express.static('public')); // enable serving files from public
 app.use(express.json()); // enable reading incoming json data
 
 //setup authentication routes
-app.use('api/auth', authRoutes);
+app.use('/api/auth', authRoutes);
 
 app.use('/api', ensureAuth);
 
@@ -52,7 +52,10 @@ app.get('/api/todos', (req, res) => {
              task,
              completed
         FROM todos
-    `)
+        WHERE user_id = $1;
+    `,
+    [req.userId]
+    )
         .then(result => {
             res.json(result.rows);
         })
@@ -66,11 +69,11 @@ app.get('/api/todos', (req, res) => {
 app.post('/api/todos', (req, res) => {
     const todo = req.body;
     client.query(`
-        INSERT INTO todos (task, completed)
-        VALUES ($1, $2)
+        INSERT INTO todos (task, completed, user_id)
+        VALUES ($1, $2, $3)
         RETURNING *;
     `,
-    [todo.task, false]
+    [todo.task, false, req.userId]
     )
         .then(result => {
             res.json(result.rows[0]);
@@ -92,9 +95,10 @@ app.put('/api/todos/:id', (req, res) => {
         SET    task = $2,
                completed = $3
         WHERE  id = $1
+        AND    user_id = $4
         RETURNING *;
     `,
-    [id, todo.task, todo.completed]
+    [id, todo.task, todo.completed, req.userId]
     )
         .then(result => {
             console.log(result);
@@ -113,9 +117,10 @@ app.delete('/api/todos/:id', (req, res) => {
     client.query(`
         DELETE FROM todos
         WHERE  id = $1
+        AND    user_id = $2
         RETURNING *;
     `,
-    [id]
+    [id, req.userId]
     )
         .then(result => {
             console.log(result);
